Some refacto
